Reorganized categorical tests
