+++ pandas/tests/categorical/test_constructors.py
@@ -6,13 +6,12 @@ import numpy as np
 
 import pandas as pd
 import pandas.util.testing as tm
-from pandas import (Categorical, Index, Series, DataFrame, Timestamp,
+from pandas import (Categorical, Index, Series, Timestamp,
                     CategoricalIndex, date_range, DatetimeIndex,
                     period_range, timedelta_range, NaT,
                     Interval, IntervalIndex)
 from pandas.core.dtypes.dtypes import CategoricalDtype
 from pandas.core.dtypes.common import is_float_dtype, is_integer_dtype
-from pandas.tests.categorical.common import TestCategoricalBlock
 
 
 class TestCategoricalConstructors(object):
@@ -508,174 +507,3 @@ class TestCategoricalConstructors(object):
         assert not cat.ordered
         cat = Categorical([0, 1, 2], ordered=True)
         assert cat.ordered
-
-
-class TestCategoricalBlockConstructorsWithFactor(TestCategoricalBlock):
-
-    def test_basic(self):
-
-        # test basic creation / coercion of categoricals
-        s = Series(self.factor, name='A')
-        assert s.dtype == 'category'
-        assert len(s) == len(self.factor)
-        str(s.values)
-        str(s)
-
-        # in a frame
-        df = DataFrame({'A': self.factor})
-        result = df['A']
-        tm.assert_series_equal(result, s)
-        result = df.iloc[:, 0]
-        tm.assert_series_equal(result, s)
-        assert len(df) == len(self.factor)
-        str(df.values)
-        str(df)
-
-        df = DataFrame({'A': s})
-        result = df['A']
-        tm.assert_series_equal(result, s)
-        assert len(df) == len(self.factor)
-        str(df.values)
-        str(df)
-
-        # multiples
-        df = DataFrame({'A': s, 'B': s, 'C': 1})
-        result1 = df['A']
-        result2 = df['B']
-        tm.assert_series_equal(result1, s)
-        tm.assert_series_equal(result2, s, check_names=False)
-        assert result2.name == 'B'
-        assert len(df) == len(self.factor)
-        str(df.values)
-        str(df)
-
-        # GH8623
-        x = DataFrame([[1, 'John P. Doe'], [2, 'Jane Dove'],
-                       [1, 'John P. Doe']],
-                      columns=['person_id', 'person_name'])
-        x['person_name'] = Categorical(x.person_name
-                                       )  # doing this breaks transform
-
-        expected = x.iloc[0].person_name
-        result = x.person_name.iloc[0]
-        assert result == expected
-
-        result = x.person_name[0]
-        assert result == expected
-
-        result = x.person_name.loc[0]
-        assert result == expected
-
-
-class TestCategoricalBlockConstructors(object):
-
-    def test_construction_series(self):
-
-        l = [1, 2, 3, 1]
-        exp = Series(l).astype('category')
-        res = Series(l, dtype='category')
-        tm.assert_series_equal(res, exp)
-
-        l = ["a", "b", "c", "a"]
-        exp = Series(l).astype('category')
-        res = Series(l, dtype='category')
-        tm.assert_series_equal(res, exp)
-
-        # insert into frame with different index
-        # GH 8076
-        index = date_range('20000101', periods=3)
-        expected = Series(Categorical(values=[np.nan, np.nan, np.nan],
-                                      categories=['a', 'b', 'c']))
-        expected.index = index
-
-        expected = DataFrame({'x': expected})
-        df = DataFrame(
-            {'x': Series(['a', 'b', 'c'], dtype='category')}, index=index)
-        tm.assert_frame_equal(df, expected)
-
-    def test_construction_frame(self):
-
-        # GH8626
-
-        # dict creation
-        df = DataFrame({'A': list('abc')}, dtype='category')
-        expected = Series(list('abc'), dtype='category', name='A')
-        tm.assert_series_equal(df['A'], expected)
-
-        # to_frame
-        s = Series(list('abc'), dtype='category')
-        result = s.to_frame()
-        expected = Series(list('abc'), dtype='category', name=0)
-        tm.assert_series_equal(result[0], expected)
-        result = s.to_frame(name='foo')
-        expected = Series(list('abc'), dtype='category', name='foo')
-        tm.assert_series_equal(result['foo'], expected)
-
-        # list-like creation
-        df = DataFrame(list('abc'), dtype='category')
-        expected = Series(list('abc'), dtype='category', name=0)
-        tm.assert_series_equal(df[0], expected)
-
-        # ndim != 1
-        df = DataFrame([Categorical(list('abc'))])
-        expected = DataFrame({0: Series(list('abc'), dtype='category')})
-        tm.assert_frame_equal(df, expected)
-
-        df = DataFrame([Categorical(list('abc')), Categorical(list('abd'))])
-        expected = DataFrame({0: Series(list('abc'), dtype='category'),
-                              1: Series(list('abd'), dtype='category')},
-                             columns=[0, 1])
-        tm.assert_frame_equal(df, expected)
-
-        # mixed
-        df = DataFrame([Categorical(list('abc')), list('def')])
-        expected = DataFrame({0: Series(list('abc'), dtype='category'),
-                              1: list('def')}, columns=[0, 1])
-        tm.assert_frame_equal(df, expected)
-
-        # invalid (shape)
-        pytest.raises(ValueError,
-                      lambda: DataFrame([Categorical(list('abc')),
-                                         Categorical(list('abdefg'))]))
-
-        # ndim > 1
-        pytest.raises(NotImplementedError,
-                      lambda: Categorical(np.array([list('abcd')])))
-
-    def test_creation_astype(self):
-        l = ["a", "b", "c", "a"]
-        s = Series(l)
-        exp = Series(Categorical(l))
-        res = s.astype('category')
-        tm.assert_series_equal(res, exp)
-
-        l = [1, 2, 3, 1]
-        s = Series(l)
-        exp = Series(Categorical(l))
-        res = s.astype('category')
-        tm.assert_series_equal(res, exp)
-
-        df = DataFrame({"cats": [1, 2, 3, 4, 5, 6],
-                        "vals": [1, 2, 3, 4, 5, 6]})
-        cats = Categorical([1, 2, 3, 4, 5, 6])
-        exp_df = DataFrame({"cats": cats, "vals": [1, 2, 3, 4, 5, 6]})
-        df["cats"] = df["cats"].astype("category")
-        tm.assert_frame_equal(exp_df, df)
-
-        df = DataFrame({"cats": ['a', 'b', 'b', 'a', 'a', 'd'],
-                        "vals": [1, 2, 3, 4, 5, 6]})
-        cats = Categorical(['a', 'b', 'b', 'a', 'a', 'd'])
-        exp_df = DataFrame({"cats": cats, "vals": [1, 2, 3, 4, 5, 6]})
-        df["cats"] = df["cats"].astype("category")
-        tm.assert_frame_equal(exp_df, df)
-
-        # with keywords
-        l = ["a", "b", "c", "a"]
-        s = Series(l)
-        exp = Series(Categorical(l, ordered=True))
-        res = s.astype(CategoricalDtype(None, ordered=True))
-        tm.assert_series_equal(res, exp)
-
-        exp = Series(Categorical(l, categories=list('abcdef'), ordered=True))
-        res = s.astype(CategoricalDtype(list('abcdef'), ordered=True))
-        tm.assert_series_equal(res, exp)
