+++ pandas/tests/categorical/test_repr.py
@@ -10,6 +10,7 @@ from pandas.tests.categorical.common import TestCategorical
 
 
 class TestCategoricalReprWithFactor(TestCategorical):
+
     def test_print(self):
         expected = ["[a, b, b, a, a, c, c, c]",
                     "Categories (3, object): [a < b < c]"]
@@ -89,28 +90,6 @@ Length: 60
 Categories (3, object): [ああああ, いいいいい, ううううううう]"""  # noqa
 
             assert _rep(c) == expected
-
-
-class TestCategoricalBlockRepr(object):
-
-    def test_repr(self):
-        a = Series(Categorical([1, 2, 3, 4]))
-        exp = u("0    1\n1    2\n2    3\n3    4\n" +
-                "dtype: category\nCategories (4, int64): [1, 2, 3, 4]")
-
-        assert exp == a.__unicode__()
-
-        a = Series(Categorical(["a", "b"] * 25))
-        exp = u("0     a\n1     b\n" + "     ..\n" + "48    a\n49    b\n" +
-                "Length: 50, dtype: category\nCategories (2, object): [a, b]")
-        with option_context("display.max_rows", 5):
-            assert exp == repr(a)
-
-        levs = list("abcdefghijklmnopqrstuvwxyz")
-        a = Series(Categorical(["a", "b"], categories=levs, ordered=True))
-        exp = u("0    a\n1    b\n" + "dtype: category\n"
-                "Categories (26, object): [a < b < c < d ... w < x < y < z]")
-        assert exp == a.__unicode__()
 
     def test_categorical_repr(self):
         c = Categorical([1, 2, 3])
@@ -383,232 +362,6 @@ Categories (20, timedelta64[ns]): [0 days 01:00:00 < 1 days 01:00:00 < 2 days 01
                                    18 days 01:00:00 < 19 days 01:00:00]"""  # noqa
 
         assert repr(c) == exp
-
-    def test_categorical_series_repr(self):
-        s = Series(Categorical([1, 2, 3]))
-        exp = """0    1
-1    2
-2    3
-dtype: category
-Categories (3, int64): [1, 2, 3]"""
-
-        assert repr(s) == exp
-
-        s = Series(Categorical(np.arange(10)))
-        exp = """0    0
-1    1
-2    2
-3    3
-4    4
-5    5
-6    6
-7    7
-8    8
-9    9
-dtype: category
-Categories (10, int64): [0, 1, 2, 3, ..., 6, 7, 8, 9]"""
-
-        assert repr(s) == exp
-
-    def test_categorical_series_repr_ordered(self):
-        s = Series(Categorical([1, 2, 3], ordered=True))
-        exp = """0    1
-1    2
-2    3
-dtype: category
-Categories (3, int64): [1 < 2 < 3]"""
-
-        assert repr(s) == exp
-
-        s = Series(Categorical(np.arange(10), ordered=True))
-        exp = """0    0
-1    1
-2    2
-3    3
-4    4
-5    5
-6    6
-7    7
-8    8
-9    9
-dtype: category
-Categories (10, int64): [0 < 1 < 2 < 3 ... 6 < 7 < 8 < 9]"""
-
-        assert repr(s) == exp
-
-    def test_categorical_series_repr_datetime(self):
-        idx = date_range('2011-01-01 09:00', freq='H', periods=5)
-        s = Series(Categorical(idx))
-        exp = """0   2011-01-01 09:00:00
-1   2011-01-01 10:00:00
-2   2011-01-01 11:00:00
-3   2011-01-01 12:00:00
-4   2011-01-01 13:00:00
-dtype: category
-Categories (5, datetime64[ns]): [2011-01-01 09:00:00, 2011-01-01 10:00:00, 2011-01-01 11:00:00,
-                                 2011-01-01 12:00:00, 2011-01-01 13:00:00]"""  # noqa
-
-        assert repr(s) == exp
-
-        idx = date_range('2011-01-01 09:00', freq='H', periods=5,
-                         tz='US/Eastern')
-        s = Series(Categorical(idx))
-        exp = """0   2011-01-01 09:00:00-05:00
-1   2011-01-01 10:00:00-05:00
-2   2011-01-01 11:00:00-05:00
-3   2011-01-01 12:00:00-05:00
-4   2011-01-01 13:00:00-05:00
-dtype: category
-Categories (5, datetime64[ns, US/Eastern]): [2011-01-01 09:00:00-05:00, 2011-01-01 10:00:00-05:00,
-                                             2011-01-01 11:00:00-05:00, 2011-01-01 12:00:00-05:00,
-                                             2011-01-01 13:00:00-05:00]"""  # noqa
-
-        assert repr(s) == exp
-
-    def test_categorical_series_repr_datetime_ordered(self):
-        idx = date_range('2011-01-01 09:00', freq='H', periods=5)
-        s = Series(Categorical(idx, ordered=True))
-        exp = """0   2011-01-01 09:00:00
-1   2011-01-01 10:00:00
-2   2011-01-01 11:00:00
-3   2011-01-01 12:00:00
-4   2011-01-01 13:00:00
-dtype: category
-Categories (5, datetime64[ns]): [2011-01-01 09:00:00 < 2011-01-01 10:00:00 < 2011-01-01 11:00:00 <
-                                 2011-01-01 12:00:00 < 2011-01-01 13:00:00]"""  # noqa
-
-        assert repr(s) == exp
-
-        idx = date_range('2011-01-01 09:00', freq='H', periods=5,
-                         tz='US/Eastern')
-        s = Series(Categorical(idx, ordered=True))
-        exp = """0   2011-01-01 09:00:00-05:00
-1   2011-01-01 10:00:00-05:00
-2   2011-01-01 11:00:00-05:00
-3   2011-01-01 12:00:00-05:00
-4   2011-01-01 13:00:00-05:00
-dtype: category
-Categories (5, datetime64[ns, US/Eastern]): [2011-01-01 09:00:00-05:00 < 2011-01-01 10:00:00-05:00 <
-                                             2011-01-01 11:00:00-05:00 < 2011-01-01 12:00:00-05:00 <
-                                             2011-01-01 13:00:00-05:00]"""  # noqa
-
-        assert repr(s) == exp
-
-    def test_categorical_series_repr_period(self):
-        idx = period_range('2011-01-01 09:00', freq='H', periods=5)
-        s = Series(Categorical(idx))
-        exp = """0   2011-01-01 09:00
-1   2011-01-01 10:00
-2   2011-01-01 11:00
-3   2011-01-01 12:00
-4   2011-01-01 13:00
-dtype: category
-Categories (5, period[H]): [2011-01-01 09:00, 2011-01-01 10:00, 2011-01-01 11:00, 2011-01-01 12:00,
-                            2011-01-01 13:00]"""  # noqa
-
-        assert repr(s) == exp
-
-        idx = period_range('2011-01', freq='M', periods=5)
-        s = Series(Categorical(idx))
-        exp = """0   2011-01
-1   2011-02
-2   2011-03
-3   2011-04
-4   2011-05
-dtype: category
-Categories (5, period[M]): [2011-01, 2011-02, 2011-03, 2011-04, 2011-05]"""
-
-        assert repr(s) == exp
-
-    def test_categorical_series_repr_period_ordered(self):
-        idx = period_range('2011-01-01 09:00', freq='H', periods=5)
-        s = Series(Categorical(idx, ordered=True))
-        exp = """0   2011-01-01 09:00
-1   2011-01-01 10:00
-2   2011-01-01 11:00
-3   2011-01-01 12:00
-4   2011-01-01 13:00
-dtype: category
-Categories (5, period[H]): [2011-01-01 09:00 < 2011-01-01 10:00 < 2011-01-01 11:00 < 2011-01-01 12:00 <
-                            2011-01-01 13:00]"""  # noqa
-
-        assert repr(s) == exp
-
-        idx = period_range('2011-01', freq='M', periods=5)
-        s = Series(Categorical(idx, ordered=True))
-        exp = """0   2011-01
-1   2011-02
-2   2011-03
-3   2011-04
-4   2011-05
-dtype: category
-Categories (5, period[M]): [2011-01 < 2011-02 < 2011-03 < 2011-04 < 2011-05]"""
-
-        assert repr(s) == exp
-
-    def test_categorical_series_repr_timedelta(self):
-        idx = timedelta_range('1 days', periods=5)
-        s = Series(Categorical(idx))
-        exp = """0   1 days
-1   2 days
-2   3 days
-3   4 days
-4   5 days
-dtype: category
-Categories (5, timedelta64[ns]): [1 days, 2 days, 3 days, 4 days, 5 days]"""
-
-        assert repr(s) == exp
-
-        idx = timedelta_range('1 hours', periods=10)
-        s = Series(Categorical(idx))
-        exp = """0   0 days 01:00:00
-1   1 days 01:00:00
-2   2 days 01:00:00
-3   3 days 01:00:00
-4   4 days 01:00:00
-5   5 days 01:00:00
-6   6 days 01:00:00
-7   7 days 01:00:00
-8   8 days 01:00:00
-9   9 days 01:00:00
-dtype: category
-Categories (10, timedelta64[ns]): [0 days 01:00:00, 1 days 01:00:00, 2 days 01:00:00,
-                                   3 days 01:00:00, ..., 6 days 01:00:00, 7 days 01:00:00,
-                                   8 days 01:00:00, 9 days 01:00:00]"""  # noqa
-
-        assert repr(s) == exp
-
-    def test_categorical_series_repr_timedelta_ordered(self):
-        idx = timedelta_range('1 days', periods=5)
-        s = Series(Categorical(idx, ordered=True))
-        exp = """0   1 days
-1   2 days
-2   3 days
-3   4 days
-4   5 days
-dtype: category
-Categories (5, timedelta64[ns]): [1 days < 2 days < 3 days < 4 days < 5 days]"""  # noqa
-
-        assert repr(s) == exp
-
-        idx = timedelta_range('1 hours', periods=10)
-        s = Series(Categorical(idx, ordered=True))
-        exp = """0   0 days 01:00:00
-1   1 days 01:00:00
-2   2 days 01:00:00
-3   3 days 01:00:00
-4   4 days 01:00:00
-5   5 days 01:00:00
-6   6 days 01:00:00
-7   7 days 01:00:00
-8   8 days 01:00:00
-9   9 days 01:00:00
-dtype: category
-Categories (10, timedelta64[ns]): [0 days 01:00:00 < 1 days 01:00:00 < 2 days 01:00:00 <
-                                   3 days 01:00:00 ... 6 days 01:00:00 < 7 days 01:00:00 <
-                                   8 days 01:00:00 < 9 days 01:00:00]"""  # noqa
-
-        assert repr(s) == exp
 
     def test_categorical_index_repr(self):
         idx = CategoricalIndex(Categorical([1, 2, 3]))
