+++ pandas/tests/frame/test_api.py
@@ -15,7 +15,7 @@ from pandas import compat
 from numpy.random import randn
 import numpy as np
 
-from pandas import DataFrame, Series, date_range, timedelta_range
+from pandas import DataFrame, Series, date_range, timedelta_range, Categorical
 import pandas as pd
 
 from pandas.util.testing import (assert_almost_equal,
@@ -239,6 +239,29 @@ class SharedWithSparse(object):
         tup3 = next(df3.itertuples())
         assert not hasattr(tup3, '_fields')
         assert isinstance(tup3, tuple)
+
+    def test_sequence_like_with_categorical(self):
+
+        # GH 7839
+        # make sure can iterate
+        df = DataFrame({"id": [1, 2, 3, 4, 5, 6],
+                        "raw_grade": ['a', 'b', 'b', 'a', 'a', 'e']})
+        df['grade'] = Categorical(df['raw_grade'])
+
+        # basic sequencing testing
+        result = list(df.grade.values)
+        expected = np.array(df.grade.values).tolist()
+        tm.assert_almost_equal(result, expected)
+
+        # iteration
+        for t in df.itertuples(index=False):
+            str(t)
+
+        for row, s in df.iterrows():
+            str(s)
+
+        for c, col in df.iteritems():
+            str(s)
 
     def test_len(self):
         assert len(self.frame) == len(self.frame.index)
